kiosks/tf_detector_service/requirements.txt
fastapi
uvicorn
numpy
opencv-python
pydantic
easyocr
inference
requests
python-dotenv



